+++ src/SerialPortWinBase.cpp
@@ -39,7 +39,7 @@ CSerialPortWinBase::CSerialPortWinBase()
     , p_buffer(new itas109::RingBuffer<char>(m_readBufferSize))
 {
     itas109::IUtils::strncpy(m_portName, "", 1);
-    m_byteReadBufferFullNotify = m_readBufferSize * 0.8;
+    m_byteReadBufferFullNotify = (unsigned int)(m_readBufferSize * 0.8);
 
     overlapMonitor.Internal = 0;
     overlapMonitor.InternalHigh = 0;
@@ -67,7 +67,7 @@ CSerialPortWinBase::CSerialPortWinBase(const char *portName)
     , p_buffer(new itas109::RingBuffer<char>(m_readBufferSize))
 {
     itas109::IUtils::strncpy(m_portName, portName, 256);
-    m_byteReadBufferFullNotify = m_readBufferSize * 0.8;
+    m_byteReadBufferFullNotify = (unsigned int)(m_readBufferSize * 0.8);
 
     overlapMonitor.Internal = 0;
     overlapMonitor.InternalHigh = 0;
@@ -102,7 +102,7 @@ void CSerialPortWinBase::init(const char *portName,
     m_stopbits = stopbits;
     m_flowControl = flowControl;
     m_readBufferSize = readBufferSize;
-    m_byteReadBufferFullNotify = m_readBufferSize * 0.8;
+    m_byteReadBufferFullNotify = (unsigned int)(m_readBufferSize * 0.8);
 
     if (p_buffer)
     {
